Create basic add book page.
